orion-platform/requirements.txt
# Core dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
asyncio-nest>=1.5.6

# AI/ML dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.30.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Database and storage
neo4j>=5.0.0
redis>=5.0.0
pymongo>=4.5.0
sqlalchemy>=2.0.0
aioredis>=2.0.0

# Knowledge graph and RDF
rdflib>=7.0.0
owlready2>=0.40
networkx>=3.0

# Materials science
pymatgen>=2023.5.0
ase>=3.22.0
rdkit>=2023.3.1

# API and web
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
aiohttp>=3.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.0
tqdm>=4.65.0
psutil>=5.9.0
gputil>=1.4.0

# Monitoring
prometheus-client>=0.17.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0